Create ALV_SORTING_TO_PERTICULAR_FIEDS
This ABAP program is an ALV report that shows the filtering concept. It allows users to filter sales order data by entering a range of sales document numbers. It fetches header data from VBAK and item data from VBAP, combines them, and displays the result in a sorted ALV grid.